+++ tests/pda-derivation/tests/typescript.spec.ts
@@ -121,4 +121,17 @@ describe("typescript", () => {
   it("Can use unsupported expressions", () => {
     // Compilation test to fix issues like https://github.com/coral-xyz/anchor/issues/2933
   });
+
+  it("Includes the unresolved accounts if resolution fails", async () => {
+    try {
+      // `unknown` account is required for account resolution to work, but it's
+      // intentionally not provided to test the error message
+      await program.methods.resolutionError().rpc();
+      throw new Error("Should throw due to account resolution failure!");
+    } catch (e) {
+      expect(e.message).to.equal(
+        "Reached maximum depth for account resolution. Unresolved accounts: `pda`, `anotherPda`"
+      );
+    }
+  });
 });